requirements.txt
# Core Agent Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# MCP Tool Servers
fastmcp>=0.1.0

# API Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Frontend
streamlit>=1.40.0

# Data & Validation
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0

# PDF Export (Bonus)
fpdf2>=2.7.0

# Utilities
httpx>=0.27.0
aiofiles>=24.0.0
python-multipart>=0.0.12

# RAG Chatbot Context
faiss-cpu>=1.7.4
langchain-community>=0.3.0
langsmith>=0.1.125
requests>=2.31.0
pyyaml>=6.0.1
